% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{enable_github_color}
\alias{enable_github_color}
\title{Enable Colors on Github Actions}
\usage{
enable_github_color(color = as.integer(256^3))
}
\arguments{
\item{color}{An integer giving the number of colors. Default 24bit.}
}
\value{
\code{TRUE} if the option was set, \code{FALSE} otherwise.
}
\description{
This will set the option \code{cli.num_colors} to \code{color}
}
\examples{
Sys.setenv(GITHUB_ACTIONS = "true")
enable_github_colors()
}
