% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{octo_add_path}
\alias{octo_add_path}
\title{Add a system path}
\usage{
octo_add_path(dir, check = TRUE)
}
\arguments{
\item{dir}{A directory.}

\item{check}{Should be checked that \code{dir} is an existing dir.}
}
\description{
Prepends a directory so the runners \code{PATH} envvar in a way that make it
available in the following steps of the action. The \code{PATH} will not update
during this step.
}
\examples{
\dontrun{
octo_addpath("/.local/bin")
}
}
\seealso{
The \href{https://github.com/assignUser/octolog/actions/workflows/test-octolog.yaml}{{octolog} example workflow}
and the \href{https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#adding-a-system-path}{Github Docs}.
}
