% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{octo_mask_value}
\alias{octo_mask_value}
\title{Masking a value in the GHA log.}
\usage{
octo_mask_value(value)
}
\arguments{
\item{value}{A single value to mask, coercible to string.}
}
\description{
This will set an envvar \verb{OCTOLOG_MASK*} containing the value and masks it.
}
\seealso{
\href{https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#masking-a-value-in-log}{Github Docs}
}
