% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{octo_set_envvar}
\alias{octo_set_envvar}
\title{Set environment variables}
\usage{
octo_set_envvar(value, name, set = TRUE, delim = "EOF")
}
\arguments{
\item{value}{Value of the envvar, coercible to string. Can be a multiline
string or character vector of \code{length > 1}, each element will be
interpreted as one line.}

\item{name}{Name of the envvar.}

\item{set}{Should the envvar also be set in this step?}

\item{delim}{Delimter used for multiline strings. No need to change this
unless your string contains 'EOF'.}
}
\description{
This will set an envrionment variable in a way that makes it available to the
following steps, compared to using \code{\link[=Sys.setenv]{Sys.setenv()}}, which would only make an
envvar available in the current step.
}
\seealso{
\code{\link[=octo_mask_envvar]{octo_mask_envvar()}}
and the \href{https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#setting-an-environment-variable}{Github docs}
}
