% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{octo_set_output}
\alias{octo_set_output}
\title{Set an output parameter}
\usage{
octo_set_output(value, name)
}
\arguments{
\item{value}{A single line string. Use \code{\link[=encode_string]{encode_string()}} to encode a numeric
or multiline string.}

\item{name}{A character vector length 1.}
}
\description{
Set an actions output parameter. These can be accessed in later steps.
}
\examples{
Sys.setenv(GITHUB_ACTION = "true")
string <- "A multi-line \n String."
value <- encode_string(string)
octo_set_output("important-value", value)
}
\seealso{
\href{https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#setting-an-output-parameter}{Github Docs}
}
