% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{octo_start_group}
\alias{octo_start_group}
\alias{octo_end_group}
\title{Grouping log lines}
\usage{
octo_start_group(title)

octo_end_group()
}
\arguments{
\item{title}{Name of the group, single line.}
}
\description{
These functions make it possible to group lines in the Github Actions log.
Groups can not be nested at this point, see this \href{https://github.com/actions/runner/issues/802}{isssue}.
}
