% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{octo_stop_commands}
\alias{octo_stop_commands}
\title{Stop workflow commands}
\usage{
octo_stop_commands()
}
\arguments{
\item{token}{A unique token used to restart workflow command parsing.}
}
\value{
The \code{token} needed to reactivate the workflow command parsing.
}
\description{
This will stop github from processing any workflow commands until
\code{\link[=octo_start_commands]{octo_start_commands()}} is called with the correct \code{token}.
}
\examples{
Sys.setenv(GITHUB_ACTION = "true")
tk <- octo_stop_commands()
# Commands will not be parsed by Github Actions
octo_start_commands(tk)
}
