% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{encode_string}
\alias{encode_string}
\title{Encode String for GitHub Actions}
\usage{
encode_string(string, join = FALSE)
}
\arguments{
\item{string}{A character vector.}

\item{join}{Join vector into single string using encoded newline.}
}
\value{
A character vector (of length 1 if \code{join}).
}
\description{
Encodes a multiline string into one line for GitHub Action output.
}
\details{
This will only encode '\%', '\\n', '\\r' as these will be automatically
decoded by GitHub when using the output via \verb{$\{\{ steps.<step_id>.outputs.<name> \}\}}. You can use \code{\link[utils:URLencode]{utils::URLencode()}} instead
of this function to also escape everything else problematic like (double)
quotes etc. but you will have to manually use \code{\link[utils:URLencode]{utils::URLdecode()}} to revert
this.
}
\examples{
chrs <- c("100\% This is some output with \n", "a new line")
encode_string(chrs)

# encode some md (e.g. to post as comment)
md <- c("# Important PR Comment", " ", "This commit is great!")
encode_string(md, TRUE)
}
