% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{octo_start_group}
\alias{octo_start_group}
\alias{octo_end_group}
\title{Grouping log lines}
\usage{
octo_start_group(name)

octo_end_group()
}
\arguments{
\item{name}{Name of the group, single line.}
}
\description{
These functions make it possible to group lines in the Github Actions log.
Groups can not be nested at this point, see this \link{issue}(https://github.com
/actions/runner/issues/802).
}
\examples{
Sys.setenv(GITHUB_ACTIONS = "TRUE")
octo_start_group("Print stuff")
print("Log other output")
octo_end_group()
}
\seealso{
\link{Github Docs}(https://docs.github.com/en/actions/using-workflows/
workflow-commands-for-github-actions#grouping-log-lines)
}
