% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{octo_stop_commands}
\alias{octo_stop_commands}
\alias{octo_start_commands}
\title{Stop workflow commands}
\usage{
octo_stop_commands()

octo_start_commands(token)
}
\arguments{
\item{token}{A unique token used to restart workflow command parsing.}
}
\value{
The \code{token} needed to reactivate the workflow command parsing.
}
\description{
This will stop github from processing any workflow commands until
\code{\link[=octo_start_commands]{octo_start_commands()}} is called with the correct \code{token}. This can be used if untrusted output (e.g. issue titles, bodies or commit messages) needs to be logged this can be used to stop this output from running possibly malicous workflow commands.
}
\examples{
Sys.setenv(GITHUB_ACTIONS = "true")
tk <- octo_stop_commands()
# Commands will not be parsed by Github Actions
octo_start_commands(tk)
}
\seealso{
The \href{https://github.com/assignUser/octolog/actions/workflows/test-octolog.yaml}{example workflow}
and the \href{https://github.blog/changelog/2020-10-01-github-actions-deprecating-set-env-and-add-path-commands/}{Github Blog}
}
