% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{octocat}
\alias{octocat}
\title{"cat" a string if on Github}
\usage{
octocat(string, last = "\\n")
}
\arguments{
\item{string}{A character vector of length 1. This string will be terminated
with \code{last} and printed with \code{\link[base:cat]{base::cat()}}.}

\item{last}{Element to terminate string. Default is \verb{\\n}.}
}
\value{
\code{string} invisibly
}
\description{
Prints a \code{string} if on Github, detected via env var \code{GITHUB_ACTIONS == 'true'}.
}
\examples{
Sys.setenv(GITHUB_ACTIONS = "true")
octocat("::error ::Something is wrong")
Sys.unsetenv("GITHUB_ACTIONS")
octocat("::error ::Something is wrong")

}
