% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octolog-package.R
\docType{package}
\name{octolog-package}
\alias{octolog}
\alias{octolog-package}
\title{octolog: Better Github Action Logging}
\description{
Octolog provides drop-in replacementsmakes it easy to create beautiful output in Github Action logs and use workflow commands in R.
}
\author{
\strong{Maintainer}: Jacob Wujciak-Jens \email{jacob@wujciak.de} (\href{https://orcid.org/0000-0002-7281-3989}{ORCID})

}
\keyword{internal}
