% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octolog-package.R
\docType{package}
\name{octolog-package}
\alias{octolog}
\alias{octolog-package}
\title{octolog: Better 'Github Actions' Logging}
\description{
Provides an API for 'Github' \url{https://github.com} workflow commands and makes it possible to signal conditions from within R that will create annotations when used within 'Github Actions' \url{https://github.com/features/actions} but raise normal R conditions when used interactively.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/assignUser/octolog}
  \item \url{https://jacob.wujciak.de/octolog/}
  \item Report bugs at \url{https://github.com/assignUser/octolog/issues}
}

}
\author{
\strong{Maintainer}: Jacob Wujciak-Jens \email{jacob@wujciak.de} (\href{https://orcid.org/0000-0002-7281-3989}{ORCID})

}
\keyword{internal}
