% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{on_github}
\alias{on_github}
\title{Detect if R is running within a GitHub Action.}
\usage{
on_github()
}
\value{
\code{TRUE} if on GitHub. \code{FALSE} otherwise.
}
\description{
Detect if R is running within a GitHub Action.
}
\details{
Uses \code{GITHUB_ACTIONS} envvar`.
}
\examples{
on_github()
}
